refactor(rust-sys): update `AsyncWasiModule`
