+++ crates/wasmedge-sys/src/instance/function.rs
@@ -11,163 +11,11 @@ use crate::{
 use core::ffi::c_void;
 use parking_lot::Mutex;
 use rand::Rng;
-#[cfg(all(feature = "async", target_os = "linux"))]
-use std::pin::Pin;
 use std::{convert::TryInto, sync::Arc};
-#[cfg(all(feature = "async", target_os = "linux"))]
-use wasmedge_types::NeverType;
 use wasmedge_types::{
     error::{FuncError, HostFuncError, WasmEdgeError},
     ValType,
 };
-
-/// Defines the signature of an asynchronous host function.
-#[cfg(all(feature = "async", target_os = "linux"))]
-pub type AsyncHostFn<T> =
-    fn(
-        CallingFrame,
-        Vec<WasmValue>,
-        Option<&'static mut T>,
-    ) -> Box<dyn std::future::Future<Output = Result<Vec<WasmValue>, HostFuncError>> + Send>;
-
-/// Defines the signature of a host function.
-#[cfg(all(feature = "async", target_os = "linux"))]
-pub type HostFn<T> = fn(
-    CallingFrame,
-    Vec<WasmValue>,
-    Option<&'static mut T>,
-) -> Result<Vec<WasmValue>, HostFuncError>;
-
-#[cfg(all(feature = "async", target_os = "linux"))]
-extern "C" fn wrap_sync_wasi_fn<T: 'static>(
-    key_ptr: *mut c_void,
-    data: *mut std::os::raw::c_void,
-    call_frame_ctx: *const ffi::WasmEdge_CallingFrameContext,
-    params: *const ffi::WasmEdge_Value,
-    param_len: u32,
-    returns: *mut ffi::WasmEdge_Value,
-    return_len: u32,
-) -> ffi::WasmEdge_Result {
-    let frame = CallingFrame::create(call_frame_ctx);
-
-    // recover the async host function
-    let real_func: HostFn<T> = unsafe { std::mem::transmute(key_ptr) };
-
-    // recover the context data
-    let data = if std::any::TypeId::of::<T>() == std::any::TypeId::of::<NeverType>() {
-        None
-    } else {
-        let data: &'static mut T = unsafe { &mut *(data as *mut T) };
-        Some(data)
-    };
-
-    // input arguments
-    let input = {
-        let raw_input = unsafe {
-            std::slice::from_raw_parts(
-                params,
-                param_len
-                    .try_into()
-                    .expect("len of params should not greater than usize"),
-            )
-        };
-        raw_input.iter().map(|r| (*r).into()).collect::<Vec<_>>()
-    };
-
-    // returns
-    let return_len = return_len
-        .try_into()
-        .expect("len of returns should not greater than usize");
-    let raw_returns = unsafe { std::slice::from_raw_parts_mut(returns, return_len) };
-
-    match real_func(frame, input, data) {
-        Ok(returns) => {
-            assert!(returns.len() == return_len, "[wasmedge-sys] check the number of returns of host function. Expected: {}, actual: {}", return_len, returns.len());
-            for (idx, wasm_value) in returns.into_iter().enumerate() {
-                raw_returns[idx] = wasm_value.as_raw();
-            }
-            ffi::WasmEdge_Result { Code: 0 }
-        }
-        Err(err) => match err {
-            HostFuncError::User(code) => unsafe {
-                ffi::WasmEdge_ResultGen(ffi::WasmEdge_ErrCategory_UserLevelError, code)
-            },
-            HostFuncError::Runtime(code) => unsafe {
-                ffi::WasmEdge_ResultGen(ffi::WasmEdge_ErrCategory_WASM, code)
-            },
-        },
-    }
-}
-
-#[cfg(all(feature = "async", target_os = "linux"))]
-extern "C" fn wrap_async_wasi_fn<T: 'static>(
-    key_ptr: *mut c_void,
-    data: *mut std::os::raw::c_void,
-    call_frame_ctx: *const ffi::WasmEdge_CallingFrameContext,
-    params: *const ffi::WasmEdge_Value,
-    param_len: u32,
-    returns: *mut ffi::WasmEdge_Value,
-    return_len: u32,
-) -> ffi::WasmEdge_Result {
-    let frame = CallingFrame::create(call_frame_ctx);
-
-    // recover the async host function
-    let real_func: AsyncHostFn<T> = unsafe { std::mem::transmute(key_ptr) };
-
-    // recover the context data
-    let data = if std::any::TypeId::of::<T>() == std::any::TypeId::of::<NeverType>() {
-        None
-    } else {
-        let data: &'static mut T = unsafe { &mut *(data as *mut T) };
-        Some(data)
-    };
-
-    // arguments
-    let input = {
-        let raw_input = unsafe {
-            std::slice::from_raw_parts(
-                params,
-                param_len
-                    .try_into()
-                    .expect("len of params should not greater than usize"),
-            )
-        };
-        raw_input.iter().map(|r| (*r).into()).collect::<Vec<_>>()
-    };
-
-    // returns
-    let return_len = return_len
-        .try_into()
-        .expect("len of returns should not greater than usize");
-    let raw_returns = unsafe { std::slice::from_raw_parts_mut(returns, return_len) };
-
-    let async_cx = AsyncCx::new();
-    let mut future = Pin::from(real_func(frame, input, data));
-    let result = match unsafe { async_cx.block_on(future.as_mut()) } {
-        Ok(Ok(ret)) => Ok(ret),
-        Ok(Err(err)) => Err(err),
-        Err(_err) => Err(HostFuncError::User(0x87)),
-    };
-
-    // parse result
-    match result {
-        Ok(returns) => {
-            assert!(returns.len() == return_len, "[wasmedge-sys] check the number of returns of async host function. Expected: {}, actual: {}", return_len, returns.len());
-            for (idx, wasm_value) in returns.into_iter().enumerate() {
-                raw_returns[idx] = wasm_value.as_raw();
-            }
-            ffi::WasmEdge_Result { Code: 0 }
-        }
-        Err(err) => match err {
-            HostFuncError::User(code) => unsafe {
-                ffi::WasmEdge_ResultGen(ffi::WasmEdge_ErrCategory_UserLevelError, code)
-            },
-            HostFuncError::Runtime(code) => unsafe {
-                ffi::WasmEdge_ResultGen(ffi::WasmEdge_ErrCategory_WASM, code)
-            },
-        },
-    }
-}
 
 pub type CustomFnWrapper = unsafe extern "C" fn(
     key_ptr: *mut c_void,
@@ -553,68 +401,6 @@ impl Function {
         }
     }
 
-    #[cfg(all(feature = "async", target_os = "linux"))]
-    pub(crate) fn create_wasi_func<T>(
-        ty: &FuncType,
-        real_fn: HostFn<T>,
-        ctx_data: Option<&mut T>,
-        cost: u64,
-    ) -> WasmEdgeResult<Self> {
-        let data = match ctx_data {
-            Some(d) => d as *mut T as *mut std::os::raw::c_void,
-            None => std::ptr::null_mut(),
-        };
-
-        let ctx = unsafe {
-            ffi::WasmEdge_FunctionInstanceCreateBinding(
-                ty.inner.0,
-                Some(wrap_sync_wasi_fn::<T>),
-                real_fn as *mut _,
-                data,
-                cost,
-            )
-        };
-
-        match ctx.is_null() {
-            true => Err(Box::new(WasmEdgeError::Func(FuncError::Create))),
-            false => Ok(Self {
-                inner: Arc::new(Mutex::new(InnerFunc(ctx))),
-                registered: false,
-            }),
-        }
-    }
-
-    #[cfg(all(feature = "async", target_os = "linux"))]
-    pub(crate) fn create_async_wasi_func<T>(
-        ty: &FuncType,
-        real_fn: AsyncHostFn<T>,
-        ctx_data: Option<&mut T>,
-        cost: u64,
-    ) -> WasmEdgeResult<Self> {
-        let data = match ctx_data {
-            Some(d) => d as *mut T as *mut std::os::raw::c_void,
-            None => std::ptr::null_mut(),
-        };
-
-        let ctx = unsafe {
-            ffi::WasmEdge_FunctionInstanceCreateBinding(
-                ty.inner.0,
-                Some(wrap_async_wasi_fn::<T>),
-                real_fn as *mut _,
-                data,
-                cost,
-            )
-        };
-
-        match ctx.is_null() {
-            true => Err(Box::new(WasmEdgeError::Func(FuncError::Create))),
-            false => Ok(Self {
-                inner: Arc::new(Mutex::new(InnerFunc(ctx))),
-                registered: false,
-            }),
-        }
-    }
-
     /// Returns the underlying wasm type of this [Function].
     ///
     /// # Errors
@@ -929,799 +715,799 @@ pub(crate) struct InnerFuncRef(pub(crate) *const ffi::WasmEdge_FunctionInstanceC
 unsafe impl Send for InnerFuncRef {}
 unsafe impl Sync for InnerFuncRef {}
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    #[cfg(all(feature = "async", target_os = "linux"))]
-    use crate::{r#async::AsyncWasiModule, ASYNC_HOST_FUNCS};
-    use crate::{
-        types::WasmValue, AsImport, Executor, ImportModule, ImportObject, Store,
-        HOST_FUNC_FOOTPRINTS,
-    };
-    use std::{
-        sync::{Arc, Mutex},
-        thread,
-    };
-    use wasmedge_macro::sys_host_function;
-    use wasmedge_types::{NeverType, ValType};
-
-    #[test]
-    fn test_func_type() {
-        // test FuncType with args and returns
-        {
-            let param_tys = vec![
-                ValType::I32,
-                ValType::I64,
-                ValType::F32,
-                ValType::F64,
-                ValType::V128,
-                ValType::ExternRef,
-            ];
-            let param_len = param_tys.len();
-            let ret_tys = vec![ValType::FuncRef, ValType::ExternRef, ValType::V128];
-            let ret_len = ret_tys.len();
-
-            // create FuncType
-            let result = FuncType::create(param_tys, ret_tys);
-            assert!(result.is_ok());
-            let func_ty = result.unwrap();
-
-            // check parameters
-            assert_eq!(func_ty.params_len(), param_len as u32);
-            let param_tys = func_ty.params_type_iter().collect::<Vec<_>>();
-            assert_eq!(
-                param_tys,
-                vec![
-                    ValType::I32,
-                    ValType::I64,
-                    ValType::F32,
-                    ValType::F64,
-                    ValType::V128,
-                    ValType::ExternRef,
-                ]
-            );
-
-            // check returns
-            assert_eq!(func_ty.returns_len(), ret_len as u32);
-            let return_tys = func_ty.returns_type_iter().collect::<Vec<_>>();
-            assert_eq!(
-                return_tys,
-                vec![ValType::FuncRef, ValType::ExternRef, ValType::V128]
-            );
-        }
-
-        // test FuncType without args and returns
-        {
-            // create FuncType
-            let result = FuncType::create([], []);
-            assert!(result.is_ok());
-            let func_ty = result.unwrap();
-
-            assert_eq!(func_ty.params_len(), 0);
-            assert_eq!(func_ty.returns_len(), 0);
-        }
-    }
-
-    #[test]
-    fn test_func_basic() {
-        #[derive(Debug)]
-        struct Data<T, S> {
-            _x: i32,
-            _y: String,
-            _v: Vec<T>,
-            _s: Vec<S>,
-        }
-        let data: Data<i32, &str> = Data {
-            _x: 12,
-            _y: "hello".to_string(),
-            _v: vec![1, 2, 3],
-            _s: vec!["macos", "linux", "windows"],
-        };
-
-        fn real_add<T: core::fmt::Debug>(
-            _frame: CallingFrame,
-            input: Vec<WasmValue>,
-            data: *mut std::ffi::c_void,
-        ) -> Result<Vec<WasmValue>, HostFuncError> {
-            println!("Rust: Entering Rust function real_add");
-
-            let host_data = unsafe { Box::from_raw(data as *mut T) };
-            println!("host_data: {:?}", host_data);
-
-            if input.len() != 2 {
-                return Err(HostFuncError::User(1));
-            }
-
-            let a = if input[0].ty() == ValType::I32 {
-                input[0].to_i32()
-            } else {
-                return Err(HostFuncError::User(2));
-            };
-
-            let b = if input[1].ty() == ValType::I32 {
-                input[1].to_i32()
-            } else {
-                return Err(HostFuncError::User(3));
-            };
-
-            let c = a + b;
-            println!("Rust: calcuating in real_add c: {c:?}");
-
-            println!("Rust: Leaving Rust function real_add");
-            Ok(vec![WasmValue::from_i32(c)])
-        }
-
-        assert_eq!(HOST_FUNCS.read().len(), 0);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
-
-        // create a FuncType
-        let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
-        assert!(result.is_ok());
-        let func_ty = result.unwrap();
-        // create a host function
-        let result = Function::create_sync_func(
-            &func_ty,
-            Box::new(real_add::<Data<i32, &str>>),
-            Some(Box::new(data)),
-            0,
-        );
-        assert!(result.is_ok());
-        let host_func = result.unwrap();
-
-        // get func type
-        let result = host_func.ty();
-        assert!(result.is_ok());
-        let ty = result.unwrap();
-
-        // check parameters
-        assert_eq!(ty.params_len(), 2);
-        let param_tys = ty.params_type_iter().collect::<Vec<_>>();
-        assert_eq!(param_tys, vec![ValType::I32; 2]);
-
-        // check returns
-        assert_eq!(ty.returns_len(), 1);
-        let return_tys = ty.returns_type_iter().collect::<Vec<_>>();
-        assert_eq!(return_tys, vec![ValType::I32]);
-
-        // run this function
-        let result = Executor::create(None, None);
-        assert!(result.is_ok());
-        let mut executor = result.unwrap();
-        let result = host_func.call(
-            &mut executor,
-            vec![WasmValue::from_i32(1), WasmValue::from_i32(2)],
-        );
-        assert!(result.is_ok());
-        let returns = result.unwrap();
-        assert_eq!(returns[0].to_i32(), 3);
-    }
-
-    #[test]
-    #[allow(clippy::assertions_on_result_states)]
-    fn test_func_create_host_func_in_host_func() {
-        #[sys_host_function]
-        fn func(
-            _frame: CallingFrame,
-            _input: Vec<WasmValue>,
-        ) -> Result<Vec<WasmValue>, HostFuncError> {
-            println!("Entering host function: func");
-
-            // spawn a new thread to create a new host function
-            let handler = std::thread::spawn(|| {
-                #[sys_host_function]
-                fn real_add(
-                    _frame: CallingFrame,
-                    input: Vec<WasmValue>,
-                ) -> Result<Vec<WasmValue>, HostFuncError> {
-                    println!("Rust: Entering Rust function real_add");
-
-                    if input.len() != 2 {
-                        return Err(HostFuncError::User(1));
-                    }
-
-                    let a = if input[0].ty() == ValType::I32 {
-                        input[0].to_i32()
-                    } else {
-                        return Err(HostFuncError::User(2));
-                    };
-
-                    let b = if input[1].ty() == ValType::I32 {
-                        input[1].to_i32()
-                    } else {
-                        return Err(HostFuncError::User(3));
-                    };
-
-                    let c = a + b;
-
-                    println!("Rust: Leaving Rust function real_add");
-                    Ok(vec![WasmValue::from_i32(c)])
-                }
-
-                // create a FuncType
-                let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
-                assert!(result.is_ok());
-                let func_ty = result.unwrap();
-                // create a host function
-                let result =
-                    Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
-                assert!(result.is_ok());
-                let host_func = result.unwrap();
-
-                // run this function
-                let result = Executor::create(None, None);
-                assert!(result.is_ok());
-                let mut executor = result.unwrap();
-                let result = host_func.call(
-                    &mut executor,
-                    vec![WasmValue::from_i32(1), WasmValue::from_i32(2)],
-                );
-                assert!(result.is_ok());
-                let returns = result.unwrap();
-                assert_eq!(returns[0].to_i32(), 3);
-            });
-            handler.join().unwrap();
-
-            println!("Leaving host function: func");
-            Ok(vec![])
-        }
-
-        // create a FuncType
-        let result = FuncType::create(vec![], vec![]);
-        assert!(result.is_ok());
-        let func_ty = result.unwrap();
-        // create a host function
-        let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(func), None, 0);
-        assert!(result.is_ok());
-        let host_func = result.unwrap();
-
-        // run this function
-        let result = Executor::create(None, None);
-        assert!(result.is_ok());
-        let mut executor = result.unwrap();
-        let result = host_func.call(&mut executor, vec![]);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_func_send() {
-        // create a FuncType
-        let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
-        assert!(result.is_ok());
-        let func_ty = result.unwrap();
-        // create a host function
-        let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
-        assert!(result.is_ok());
-        let host_func = result.unwrap();
-
-        let handle = thread::spawn(move || {
-            // get func type
-            let result = host_func.ty();
-            assert!(result.is_ok());
-            let ty = result.unwrap();
-
-            // check parameters
-            assert_eq!(ty.params_len(), 2);
-            let param_tys = ty.params_type_iter().collect::<Vec<_>>();
-            assert_eq!(param_tys, vec![ValType::I32; 2]);
-
-            // check returns
-            assert_eq!(ty.returns_len(), 1);
-            let return_tys = ty.returns_type_iter().collect::<Vec<_>>();
-            assert_eq!(return_tys, vec![ValType::I32]);
-        });
-
-        handle.join().unwrap()
-    }
-
-    #[test]
-    fn test_func_sync() {
-        // create a FuncType
-        let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
-        assert!(result.is_ok());
-        let func_ty = result.unwrap();
-        // create a host function
-        let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
-        assert!(result.is_ok());
-        let host_func = Arc::new(Mutex::new(result.unwrap()));
-
-        let host_func_cloned = Arc::clone(&host_func);
-        let handle = thread::spawn(move || {
-            let result = host_func_cloned.lock();
-            assert!(result.is_ok());
-            let host_func = result.unwrap();
-
-            // get func type
-            let result = host_func.ty();
-            assert!(result.is_ok());
-            let ty = result.unwrap();
-
-            // check parameters
-            assert_eq!(ty.params_len(), 2);
-            let param_tys = ty.params_type_iter().collect::<Vec<_>>();
-            assert_eq!(param_tys, vec![ValType::I32; 2]);
-
-            // check returns
-            assert_eq!(ty.returns_len(), 1);
-            let return_tys = ty.returns_type_iter().collect::<Vec<_>>();
-            assert_eq!(return_tys, vec![ValType::I32]);
-        });
-
-        handle.join().unwrap();
-    }
-
-    #[sys_host_function]
-    fn real_add(
-        _frame: CallingFrame,
-        input: Vec<WasmValue>,
-    ) -> Result<Vec<WasmValue>, HostFuncError> {
-        println!("Rust: Entering Rust function real_add");
-
-        if input.len() != 2 {
-            return Err(HostFuncError::User(1));
-        }
-
-        let a = if input[0].ty() == ValType::I32 {
-            input[0].to_i32()
-        } else {
-            return Err(HostFuncError::User(2));
-        };
-
-        let b = if input[1].ty() == ValType::I32 {
-            input[1].to_i32()
-        } else {
-            return Err(HostFuncError::User(3));
-        };
-
-        let c = a + b;
-        println!("Rust: calcuating in real_add c: {c:?}");
-
-        println!("Rust: Leaving Rust function real_add");
-        Ok(vec![WasmValue::from_i32(c)])
-    }
-
-    #[test]
-    fn test_func_closure() -> Result<(), Box<dyn std::error::Error>> {
-        {
-            // create a host function
-            let real_add = |_: CallingFrame,
-                            input: Vec<WasmValue>,
-                            _: *mut std::os::raw::c_void|
-             -> Result<Vec<WasmValue>, HostFuncError> {
-                println!("Rust: Entering Rust function real_add");
-
-                if input.len() != 2 {
-                    return Err(HostFuncError::User(1));
-                }
-
-                let a = if input[0].ty() == ValType::I32 {
-                    input[0].to_i32()
-                } else {
-                    return Err(HostFuncError::User(2));
-                };
-
-                let b = if input[1].ty() == ValType::I32 {
-                    input[1].to_i32()
-                } else {
-                    return Err(HostFuncError::User(3));
-                };
-
-                let c = a + b;
-                println!("Rust: calcuating in real_add c: {c:?}");
-
-                println!("Rust: Leaving Rust function real_add");
-                Ok(vec![WasmValue::from_i32(c)])
-            };
-
-            // create a FuncType
-            let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
-            assert!(result.is_ok());
-            let func_ty = result.unwrap();
-
-            // create a host function from the closure defined above
-            let result =
-                Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
-            assert!(result.is_ok());
-            let host_func = result.unwrap();
-
-            // create a Store
-            let result = Store::create();
-            assert!(result.is_ok());
-            let mut store = result.unwrap();
-
-            // create an ImportModule
-            let mut import = ImportModule::create::<NeverType>("extern", None)?;
-            import.add_func("add", host_func);
-            let extern_import = ImportObject::Import(import);
-
-            // run this function
-            let result = Executor::create(None, None);
-            assert!(result.is_ok());
-            let mut executor = result.unwrap();
-            executor.register_import_object(&mut store, &extern_import)?;
-
-            let extern_instance = store.module("extern")?;
-            let add = extern_instance.get_func("add")?;
-
-            let result =
-                executor.call_func(&add, vec![WasmValue::from_i32(1), WasmValue::from_i32(2)]);
-            assert!(result.is_ok());
-            let returns = result.unwrap();
-            assert_eq!(returns[0].to_i32(), 3);
-        }
-
-        assert_eq!(HOST_FUNCS.read().len(), 0);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_func_drop() -> Result<(), Box<dyn std::error::Error>> {
-        // create a host function
-        let real_add = |_: CallingFrame,
-                        input: Vec<WasmValue>,
-                        _: *mut std::os::raw::c_void|
-         -> Result<Vec<WasmValue>, HostFuncError> {
-            println!("Rust: Entering Rust function real_add");
-
-            if input.len() != 2 {
-                return Err(HostFuncError::User(1));
-            }
-
-            let a = if input[0].ty() == ValType::I32 {
-                input[0].to_i32()
-            } else {
-                return Err(HostFuncError::User(2));
-            };
-
-            let b = if input[1].ty() == ValType::I32 {
-                input[1].to_i32()
-            } else {
-                return Err(HostFuncError::User(3));
-            };
-
-            let c = a + b;
-            println!("Rust: calcuating in real_add c: {c:?}");
-
-            println!("Rust: Leaving Rust function real_add");
-            Ok(vec![WasmValue::from_i32(c)])
-        };
-
-        // create a FuncType
-        let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
-        assert!(result.is_ok());
-        let func_ty = result.unwrap();
-
-        // create a host function
-        let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
-        assert!(result.is_ok());
-        let host_func = result.unwrap();
-
-        assert_eq!(Arc::strong_count(&host_func.inner), 1);
-        assert!(!host_func.registered);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        // clone the host function before adding it to the import object
-        let host_func_cloned = host_func.clone();
-
-        assert_eq!(Arc::strong_count(&host_func_cloned.inner), 2);
-        assert!(!host_func_cloned.registered);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        // create an ImportModule
-        let mut import_module = ImportModule::create::<NeverType>("extern", None)?;
-        // add the host function to the import module
-        import_module.add_func("add", host_func);
-
-        assert_eq!(Arc::strong_count(&host_func_cloned.inner), 2);
-        assert!(!host_func_cloned.registered);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        drop(host_func_cloned);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        let import = ImportObject::Import(import_module);
-
-        // create a Store
-        let result = Store::create();
-        assert!(result.is_ok());
-        let mut store = result.unwrap();
-
-        // run this function
-        let result = Executor::create(None, None);
-        assert!(result.is_ok());
-        let mut executor = result.unwrap();
-        executor.register_import_object(&mut store, &import)?;
-
-        // get the registered host function
-        let extern_instance = store.module("extern")?;
-        let add = extern_instance.get_func("add")?;
-        assert_eq!(Arc::strong_count(&add.inner), 1);
-        assert!(add.registered);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        // clone the host function
-        let add_cloned = add.clone();
-        assert_eq!(Arc::strong_count(&add.inner), 2);
-        assert!(add.registered);
-        assert_eq!(Arc::strong_count(&add_cloned.inner), 2);
-        assert!(add_cloned.registered);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        // drop the cloned host function
-        drop(add_cloned);
-        assert_eq!(Arc::strong_count(&add.inner), 1);
-        assert!(add.registered);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        drop(add);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        // get the registered host function again
-        let extern_instance = store.module("extern")?;
-        let add_again = extern_instance.get_func("add")?;
-        assert_eq!(Arc::strong_count(&add_again.inner), 1);
-        assert!(add_again.registered);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        // ! notice that `add_again` should be dropped before or not be used after dropping `import`
-        drop(add_again);
-
-        assert_eq!(HOST_FUNCS.read().len(), 1);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
-
-        // drop the import object
-        drop(import);
-
-        assert!(store.module("extern").is_err());
-
-        assert_eq!(HOST_FUNCS.read().len(), 0);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
-
-        // ! if `add_again` is not dropped before dropping `import`, then calling `add_again` will crash
-        // let result = executor.call_func(
-        //     &add_again,
-        //     vec![WasmValue::from_i32(1), WasmValue::from_i32(2)],
-        // );
-
-        Ok(())
-    }
-
-    #[cfg(all(feature = "async", target_os = "linux"))]
-    #[tokio::test]
-    async fn test_func_async_closure() -> Result<(), Box<dyn std::error::Error>> {
-        {
-            #[derive(Debug)]
-            struct Data<T, S> {
-                _x: i32,
-                _y: String,
-                _v: Vec<T>,
-                _s: Vec<S>,
-            }
-            impl<T, S> Drop for Data<T, S> {
-                fn drop(&mut self) {
-                    println!("Dropping Data");
-                }
-            }
-
-            let data: Data<i32, &str> = Data {
-                _x: 12,
-                _y: "hello".to_string(),
-                _v: vec![1, 2, 3],
-                _s: vec!["macos", "linux", "windows"],
-            };
-
-            // define an async closure
-            let c = |_frame: CallingFrame,
-                     _args: Vec<WasmValue>,
-                     data: *mut std::os::raw::c_void|
-             -> Box<
-                (dyn std::future::Future<Output = Result<Vec<WasmValue>, HostFuncError>> + Send),
-            > {
-                // let host_data = unsafe { &mut *(data as *mut Data<i32, &str>) };
-                let host_data = unsafe { Box::from_raw(data as *mut Data<i32, &str>) };
-
-                Box::new(async move {
-                    for _ in 0..10 {
-                        println!("[async hello] say hello");
-                        tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-
-                        println!("host_data: {:?}", host_data);
-                    }
-
-                    println!("[async hello] Done!");
-
-                    Ok(vec![])
-                })
-            };
-
-            // create a FuncType
-            let result = FuncType::create(vec![], vec![]);
-            assert!(result.is_ok());
-            let func_ty = result.unwrap();
-
-            // create an async host function
-            let result =
-                Function::create_async_func(&func_ty, Box::new(c), Some(Box::new(data)), 0);
-            assert!(result.is_ok());
-            let async_hello_func = result.unwrap();
-
-            // create an Executor
-            let result = Executor::create(None, None);
-            assert!(result.is_ok());
-            let mut executor = result.unwrap();
-            assert!(!executor.inner.0.is_null());
-
-            // create a Store
-            let result = Store::create();
-            assert!(result.is_ok());
-            let mut store = result.unwrap();
-
-            // create an AsyncWasiModule
-            let result = AsyncWasiModule::create(Some(vec!["abc"]), Some(vec![("a", "1")]), None);
-            assert!(result.is_ok());
-            let async_wasi_module = result.unwrap();
-
-            // register async_wasi module into the store
-            let wasi_import = ImportObject::AsyncWasi(async_wasi_module);
-            let result = executor.register_import_object(&mut store, &wasi_import);
-            assert!(result.is_ok());
-
-            // create an ImportModule
-            let mut import = ImportModule::create::<NeverType>("extern", None)?;
-            import.add_func("async_hello", async_hello_func);
-
-            let extern_import = ImportObject::Import(import);
-            executor.register_import_object(&mut store, &extern_import)?;
-
-            let extern_instance = store.module("extern")?;
-            let async_hello = extern_instance.get_func("async_hello")?;
-
-            async fn tick() {
-                let mut i = 0;
-                loop {
-                    println!("[tick] i={i}");
-                    tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-                    i += 1;
-                }
-            }
-            tokio::spawn(tick());
-
-            let async_state = AsyncState::new();
-            let _ = executor
-                .call_func_async(&async_state, &async_hello, [])
-                .await?;
-        }
-
-        assert_eq!(ASYNC_HOST_FUNCS.read().len(), 0);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
-
-        Ok(())
-    }
-
-    #[cfg(all(feature = "async", target_os = "linux"))]
-    #[tokio::test]
-    async fn test_func_async_func() -> Result<(), Box<dyn std::error::Error>> {
-        {
-            #[derive(Debug)]
-            struct Data<T, S> {
-                _x: i32,
-                _y: String,
-                _v: Vec<T>,
-                _s: Vec<S>,
-            }
-            let data: Data<i32, &str> = Data {
-                _x: 12,
-                _y: "hello".to_string(),
-                _v: vec![1, 2, 3],
-                _s: vec!["macos", "linux", "windows"],
-            };
-
-            // define async host function
-            fn f<T: core::fmt::Debug + Send + Sync + 'static>(
-                _frame: CallingFrame,
-                _args: Vec<WasmValue>,
-                data: *mut std::ffi::c_void,
-            ) -> Box<(dyn std::future::Future<Output = Result<Vec<WasmValue>, HostFuncError>> + Send)>
-            {
-                let data = unsafe { Box::from_raw(data as *mut T) };
-
-                Box::new(async move {
-                    for _ in 0..10 {
-                        println!("[async hello] say hello");
-                        tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-                        println!("host_data: {:?}", data);
-                    }
-
-                    println!("[async hello] Done!");
-
-                    Ok(vec![])
-                })
-            }
-
-            // create a FuncType
-            let result = FuncType::create(vec![], vec![]);
-            assert!(result.is_ok());
-            let func_ty = result.unwrap();
-
-            // create an async host function
-            let result = Function::create_async_func(
-                &func_ty,
-                Box::new(f::<Data<i32, &str>>),
-                Some(Box::new(data)),
-                0,
-            );
-            assert!(result.is_ok());
-            let async_hello_func = result.unwrap();
-
-            // create an Executor
-            let result = Executor::create(None, None);
-            assert!(result.is_ok());
-            let mut executor = result.unwrap();
-            assert!(!executor.inner.0.is_null());
-
-            // create a Store
-            let result = Store::create();
-            assert!(result.is_ok());
-            let mut store = result.unwrap();
-
-            // create an AsyncWasiModule
-            let result = AsyncWasiModule::create(Some(vec!["abc"]), Some(vec![("a", "1")]), None);
-            assert!(result.is_ok());
-            let async_wasi_module = result.unwrap();
-
-            // register async_wasi module into the store
-            let wasi_import = ImportObject::AsyncWasi(async_wasi_module);
-            let result = executor.register_import_object(&mut store, &wasi_import);
-            assert!(result.is_ok());
-
-            // create an ImportModule
-            let mut import = ImportModule::create::<NeverType>("extern", None)?;
-            import.add_func("async_hello", async_hello_func);
-
-            let extern_import = ImportObject::Import(import);
-            executor.register_import_object(&mut store, &extern_import)?;
-
-            let extern_instance = store.module("extern")?;
-            let async_hello = extern_instance.get_func("async_hello")?;
-
-            async fn tick() {
-                let mut i = 0;
-                loop {
-                    println!("[tick] i={i}");
-                    tokio::time::sleep(std::time::Duration::from_millis(500)).await;
-                    i += 1;
-                }
-            }
-            tokio::spawn(tick());
-
-            let async_state = AsyncState::new();
-            let _ = executor
-                .call_func_async(&async_state, &async_hello, [])
-                .await?;
-        }
-
-        assert_eq!(ASYNC_HOST_FUNCS.read().len(), 0);
-        assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
-
-        Ok(())
-    }
-}
+// #[cfg(test)]
+// mod tests {
+//     use super::*;
+//     #[cfg(all(feature = "async", target_os = "linux"))]
+//     use crate::{r#async::AsyncWasiModule, ASYNC_HOST_FUNCS};
+//     use crate::{
+//         types::WasmValue, AsImport, Executor, ImportModule, ImportObject, Store,
+//         HOST_FUNC_FOOTPRINTS,
+//     };
+//     use std::{
+//         sync::{Arc, Mutex},
+//         thread,
+//     };
+//     use wasmedge_macro::sys_host_function;
+//     use wasmedge_types::{NeverType, ValType};
+
+//     #[test]
+//     fn test_func_type() {
+//         // test FuncType with args and returns
+//         {
+//             let param_tys = vec![
+//                 ValType::I32,
+//                 ValType::I64,
+//                 ValType::F32,
+//                 ValType::F64,
+//                 ValType::V128,
+//                 ValType::ExternRef,
+//             ];
+//             let param_len = param_tys.len();
+//             let ret_tys = vec![ValType::FuncRef, ValType::ExternRef, ValType::V128];
+//             let ret_len = ret_tys.len();
+
+//             // create FuncType
+//             let result = FuncType::create(param_tys, ret_tys);
+//             assert!(result.is_ok());
+//             let func_ty = result.unwrap();
+
+//             // check parameters
+//             assert_eq!(func_ty.params_len(), param_len as u32);
+//             let param_tys = func_ty.params_type_iter().collect::<Vec<_>>();
+//             assert_eq!(
+//                 param_tys,
+//                 vec![
+//                     ValType::I32,
+//                     ValType::I64,
+//                     ValType::F32,
+//                     ValType::F64,
+//                     ValType::V128,
+//                     ValType::ExternRef,
+//                 ]
+//             );
+
+//             // check returns
+//             assert_eq!(func_ty.returns_len(), ret_len as u32);
+//             let return_tys = func_ty.returns_type_iter().collect::<Vec<_>>();
+//             assert_eq!(
+//                 return_tys,
+//                 vec![ValType::FuncRef, ValType::ExternRef, ValType::V128]
+//             );
+//         }
+
+//         // test FuncType without args and returns
+//         {
+//             // create FuncType
+//             let result = FuncType::create([], []);
+//             assert!(result.is_ok());
+//             let func_ty = result.unwrap();
+
+//             assert_eq!(func_ty.params_len(), 0);
+//             assert_eq!(func_ty.returns_len(), 0);
+//         }
+//     }
+
+//     #[test]
+//     fn test_func_basic() {
+//         #[derive(Debug)]
+//         struct Data<T, S> {
+//             _x: i32,
+//             _y: String,
+//             _v: Vec<T>,
+//             _s: Vec<S>,
+//         }
+//         let data: Data<i32, &str> = Data {
+//             _x: 12,
+//             _y: "hello".to_string(),
+//             _v: vec![1, 2, 3],
+//             _s: vec!["macos", "linux", "windows"],
+//         };
+
+//         fn real_add<T: core::fmt::Debug>(
+//             _frame: CallingFrame,
+//             input: Vec<WasmValue>,
+//             data: *mut std::ffi::c_void,
+//         ) -> Result<Vec<WasmValue>, HostFuncError> {
+//             println!("Rust: Entering Rust function real_add");
+
+//             let host_data = unsafe { Box::from_raw(data as *mut T) };
+//             println!("host_data: {:?}", host_data);
+
+//             if input.len() != 2 {
+//                 return Err(HostFuncError::User(1));
+//             }
+
+//             let a = if input[0].ty() == ValType::I32 {
+//                 input[0].to_i32()
+//             } else {
+//                 return Err(HostFuncError::User(2));
+//             };
+
+//             let b = if input[1].ty() == ValType::I32 {
+//                 input[1].to_i32()
+//             } else {
+//                 return Err(HostFuncError::User(3));
+//             };
+
+//             let c = a + b;
+//             println!("Rust: calcuating in real_add c: {c:?}");
+
+//             println!("Rust: Leaving Rust function real_add");
+//             Ok(vec![WasmValue::from_i32(c)])
+//         }
+
+//         assert_eq!(HOST_FUNCS.read().len(), 0);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
+
+//         // create a FuncType
+//         let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
+//         assert!(result.is_ok());
+//         let func_ty = result.unwrap();
+//         // create a host function
+//         let result = Function::create_sync_func(
+//             &func_ty,
+//             Box::new(real_add::<Data<i32, &str>>),
+//             Some(Box::new(data)),
+//             0,
+//         );
+//         assert!(result.is_ok());
+//         let host_func = result.unwrap();
+
+//         // get func type
+//         let result = host_func.ty();
+//         assert!(result.is_ok());
+//         let ty = result.unwrap();
+
+//         // check parameters
+//         assert_eq!(ty.params_len(), 2);
+//         let param_tys = ty.params_type_iter().collect::<Vec<_>>();
+//         assert_eq!(param_tys, vec![ValType::I32; 2]);
+
+//         // check returns
+//         assert_eq!(ty.returns_len(), 1);
+//         let return_tys = ty.returns_type_iter().collect::<Vec<_>>();
+//         assert_eq!(return_tys, vec![ValType::I32]);
+
+//         // run this function
+//         let result = Executor::create(None, None);
+//         assert!(result.is_ok());
+//         let mut executor = result.unwrap();
+//         let result = host_func.call(
+//             &mut executor,
+//             vec![WasmValue::from_i32(1), WasmValue::from_i32(2)],
+//         );
+//         assert!(result.is_ok());
+//         let returns = result.unwrap();
+//         assert_eq!(returns[0].to_i32(), 3);
+//     }
+
+//     #[test]
+//     #[allow(clippy::assertions_on_result_states)]
+//     fn test_func_create_host_func_in_host_func() {
+//         #[sys_host_function]
+//         fn func(
+//             _frame: CallingFrame,
+//             _input: Vec<WasmValue>,
+//         ) -> Result<Vec<WasmValue>, HostFuncError> {
+//             println!("Entering host function: func");
+
+//             // spawn a new thread to create a new host function
+//             let handler = std::thread::spawn(|| {
+//                 #[sys_host_function]
+//                 fn real_add(
+//                     _frame: CallingFrame,
+//                     input: Vec<WasmValue>,
+//                 ) -> Result<Vec<WasmValue>, HostFuncError> {
+//                     println!("Rust: Entering Rust function real_add");
+
+//                     if input.len() != 2 {
+//                         return Err(HostFuncError::User(1));
+//                     }
+
+//                     let a = if input[0].ty() == ValType::I32 {
+//                         input[0].to_i32()
+//                     } else {
+//                         return Err(HostFuncError::User(2));
+//                     };
+
+//                     let b = if input[1].ty() == ValType::I32 {
+//                         input[1].to_i32()
+//                     } else {
+//                         return Err(HostFuncError::User(3));
+//                     };
+
+//                     let c = a + b;
+
+//                     println!("Rust: Leaving Rust function real_add");
+//                     Ok(vec![WasmValue::from_i32(c)])
+//                 }
+
+//                 // create a FuncType
+//                 let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
+//                 assert!(result.is_ok());
+//                 let func_ty = result.unwrap();
+//                 // create a host function
+//                 let result =
+//                     Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
+//                 assert!(result.is_ok());
+//                 let host_func = result.unwrap();
+
+//                 // run this function
+//                 let result = Executor::create(None, None);
+//                 assert!(result.is_ok());
+//                 let mut executor = result.unwrap();
+//                 let result = host_func.call(
+//                     &mut executor,
+//                     vec![WasmValue::from_i32(1), WasmValue::from_i32(2)],
+//                 );
+//                 assert!(result.is_ok());
+//                 let returns = result.unwrap();
+//                 assert_eq!(returns[0].to_i32(), 3);
+//             });
+//             handler.join().unwrap();
+
+//             println!("Leaving host function: func");
+//             Ok(vec![])
+//         }
+
+//         // create a FuncType
+//         let result = FuncType::create(vec![], vec![]);
+//         assert!(result.is_ok());
+//         let func_ty = result.unwrap();
+//         // create a host function
+//         let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(func), None, 0);
+//         assert!(result.is_ok());
+//         let host_func = result.unwrap();
+
+//         // run this function
+//         let result = Executor::create(None, None);
+//         assert!(result.is_ok());
+//         let mut executor = result.unwrap();
+//         let result = host_func.call(&mut executor, vec![]);
+//         assert!(result.is_ok());
+//     }
+
+//     #[test]
+//     fn test_func_send() {
+//         // create a FuncType
+//         let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
+//         assert!(result.is_ok());
+//         let func_ty = result.unwrap();
+//         // create a host function
+//         let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
+//         assert!(result.is_ok());
+//         let host_func = result.unwrap();
+
+//         let handle = thread::spawn(move || {
+//             // get func type
+//             let result = host_func.ty();
+//             assert!(result.is_ok());
+//             let ty = result.unwrap();
+
+//             // check parameters
+//             assert_eq!(ty.params_len(), 2);
+//             let param_tys = ty.params_type_iter().collect::<Vec<_>>();
+//             assert_eq!(param_tys, vec![ValType::I32; 2]);
+
+//             // check returns
+//             assert_eq!(ty.returns_len(), 1);
+//             let return_tys = ty.returns_type_iter().collect::<Vec<_>>();
+//             assert_eq!(return_tys, vec![ValType::I32]);
+//         });
+
+//         handle.join().unwrap()
+//     }
+
+//     #[test]
+//     fn test_func_sync() {
+//         // create a FuncType
+//         let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
+//         assert!(result.is_ok());
+//         let func_ty = result.unwrap();
+//         // create a host function
+//         let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
+//         assert!(result.is_ok());
+//         let host_func = Arc::new(Mutex::new(result.unwrap()));
+
+//         let host_func_cloned = Arc::clone(&host_func);
+//         let handle = thread::spawn(move || {
+//             let result = host_func_cloned.lock();
+//             assert!(result.is_ok());
+//             let host_func = result.unwrap();
+
+//             // get func type
+//             let result = host_func.ty();
+//             assert!(result.is_ok());
+//             let ty = result.unwrap();
+
+//             // check parameters
+//             assert_eq!(ty.params_len(), 2);
+//             let param_tys = ty.params_type_iter().collect::<Vec<_>>();
+//             assert_eq!(param_tys, vec![ValType::I32; 2]);
+
+//             // check returns
+//             assert_eq!(ty.returns_len(), 1);
+//             let return_tys = ty.returns_type_iter().collect::<Vec<_>>();
+//             assert_eq!(return_tys, vec![ValType::I32]);
+//         });
+
+//         handle.join().unwrap();
+//     }
+
+//     #[sys_host_function]
+//     fn real_add(
+//         _frame: CallingFrame,
+//         input: Vec<WasmValue>,
+//     ) -> Result<Vec<WasmValue>, HostFuncError> {
+//         println!("Rust: Entering Rust function real_add");
+
+//         if input.len() != 2 {
+//             return Err(HostFuncError::User(1));
+//         }
+
+//         let a = if input[0].ty() == ValType::I32 {
+//             input[0].to_i32()
+//         } else {
+//             return Err(HostFuncError::User(2));
+//         };
+
+//         let b = if input[1].ty() == ValType::I32 {
+//             input[1].to_i32()
+//         } else {
+//             return Err(HostFuncError::User(3));
+//         };
+
+//         let c = a + b;
+//         println!("Rust: calcuating in real_add c: {c:?}");
+
+//         println!("Rust: Leaving Rust function real_add");
+//         Ok(vec![WasmValue::from_i32(c)])
+//     }
+
+//     #[test]
+//     fn test_func_closure() -> Result<(), Box<dyn std::error::Error>> {
+//         {
+//             // create a host function
+//             let real_add = |_: CallingFrame,
+//                             input: Vec<WasmValue>,
+//                             _: *mut std::os::raw::c_void|
+//              -> Result<Vec<WasmValue>, HostFuncError> {
+//                 println!("Rust: Entering Rust function real_add");
+
+//                 if input.len() != 2 {
+//                     return Err(HostFuncError::User(1));
+//                 }
+
+//                 let a = if input[0].ty() == ValType::I32 {
+//                     input[0].to_i32()
+//                 } else {
+//                     return Err(HostFuncError::User(2));
+//                 };
+
+//                 let b = if input[1].ty() == ValType::I32 {
+//                     input[1].to_i32()
+//                 } else {
+//                     return Err(HostFuncError::User(3));
+//                 };
+
+//                 let c = a + b;
+//                 println!("Rust: calcuating in real_add c: {c:?}");
+
+//                 println!("Rust: Leaving Rust function real_add");
+//                 Ok(vec![WasmValue::from_i32(c)])
+//             };
+
+//             // create a FuncType
+//             let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
+//             assert!(result.is_ok());
+//             let func_ty = result.unwrap();
+
+//             // create a host function from the closure defined above
+//             let result =
+//                 Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
+//             assert!(result.is_ok());
+//             let host_func = result.unwrap();
+
+//             // create a Store
+//             let result = Store::create();
+//             assert!(result.is_ok());
+//             let mut store = result.unwrap();
+
+//             // create an ImportModule
+//             let mut import = ImportModule::create::<NeverType>("extern", None)?;
+//             import.add_func("add", host_func);
+//             let extern_import = ImportObject::Import(import);
+
+//             // run this function
+//             let result = Executor::create(None, None);
+//             assert!(result.is_ok());
+//             let mut executor = result.unwrap();
+//             executor.register_import_object(&mut store, &extern_import)?;
+
+//             let extern_instance = store.module("extern")?;
+//             let add = extern_instance.get_func("add")?;
+
+//             let result =
+//                 executor.call_func(&add, vec![WasmValue::from_i32(1), WasmValue::from_i32(2)]);
+//             assert!(result.is_ok());
+//             let returns = result.unwrap();
+//             assert_eq!(returns[0].to_i32(), 3);
+//         }
+
+//         assert_eq!(HOST_FUNCS.read().len(), 0);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
+
+//         Ok(())
+//     }
+
+//     #[test]
+//     fn test_func_drop() -> Result<(), Box<dyn std::error::Error>> {
+//         // create a host function
+//         let real_add = |_: CallingFrame,
+//                         input: Vec<WasmValue>,
+//                         _: *mut std::os::raw::c_void|
+//          -> Result<Vec<WasmValue>, HostFuncError> {
+//             println!("Rust: Entering Rust function real_add");
+
+//             if input.len() != 2 {
+//                 return Err(HostFuncError::User(1));
+//             }
+
+//             let a = if input[0].ty() == ValType::I32 {
+//                 input[0].to_i32()
+//             } else {
+//                 return Err(HostFuncError::User(2));
+//             };
+
+//             let b = if input[1].ty() == ValType::I32 {
+//                 input[1].to_i32()
+//             } else {
+//                 return Err(HostFuncError::User(3));
+//             };
+
+//             let c = a + b;
+//             println!("Rust: calcuating in real_add c: {c:?}");
+
+//             println!("Rust: Leaving Rust function real_add");
+//             Ok(vec![WasmValue::from_i32(c)])
+//         };
+
+//         // create a FuncType
+//         let result = FuncType::create(vec![ValType::I32; 2], vec![ValType::I32]);
+//         assert!(result.is_ok());
+//         let func_ty = result.unwrap();
+
+//         // create a host function
+//         let result = Function::create_sync_func::<NeverType>(&func_ty, Box::new(real_add), None, 0);
+//         assert!(result.is_ok());
+//         let host_func = result.unwrap();
+
+//         assert_eq!(Arc::strong_count(&host_func.inner), 1);
+//         assert!(!host_func.registered);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         // clone the host function before adding it to the import object
+//         let host_func_cloned = host_func.clone();
+
+//         assert_eq!(Arc::strong_count(&host_func_cloned.inner), 2);
+//         assert!(!host_func_cloned.registered);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         // create an ImportModule
+//         let mut import_module = ImportModule::create::<NeverType>("extern", None)?;
+//         // add the host function to the import module
+//         import_module.add_func("add", host_func);
+
+//         assert_eq!(Arc::strong_count(&host_func_cloned.inner), 2);
+//         assert!(!host_func_cloned.registered);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         drop(host_func_cloned);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         let import = ImportObject::Import(import_module);
+
+//         // create a Store
+//         let result = Store::create();
+//         assert!(result.is_ok());
+//         let mut store = result.unwrap();
+
+//         // run this function
+//         let result = Executor::create(None, None);
+//         assert!(result.is_ok());
+//         let mut executor = result.unwrap();
+//         executor.register_import_object(&mut store, &import)?;
+
+//         // get the registered host function
+//         let extern_instance = store.module("extern")?;
+//         let add = extern_instance.get_func("add")?;
+//         assert_eq!(Arc::strong_count(&add.inner), 1);
+//         assert!(add.registered);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         // clone the host function
+//         let add_cloned = add.clone();
+//         assert_eq!(Arc::strong_count(&add.inner), 2);
+//         assert!(add.registered);
+//         assert_eq!(Arc::strong_count(&add_cloned.inner), 2);
+//         assert!(add_cloned.registered);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         // drop the cloned host function
+//         drop(add_cloned);
+//         assert_eq!(Arc::strong_count(&add.inner), 1);
+//         assert!(add.registered);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         drop(add);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         // get the registered host function again
+//         let extern_instance = store.module("extern")?;
+//         let add_again = extern_instance.get_func("add")?;
+//         assert_eq!(Arc::strong_count(&add_again.inner), 1);
+//         assert!(add_again.registered);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         // ! notice that `add_again` should be dropped before or not be used after dropping `import`
+//         drop(add_again);
+
+//         assert_eq!(HOST_FUNCS.read().len(), 1);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 1);
+
+//         // drop the import object
+//         drop(import);
+
+//         assert!(store.module("extern").is_err());
+
+//         assert_eq!(HOST_FUNCS.read().len(), 0);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
+
+//         // ! if `add_again` is not dropped before dropping `import`, then calling `add_again` will crash
+//         // let result = executor.call_func(
+//         //     &add_again,
+//         //     vec![WasmValue::from_i32(1), WasmValue::from_i32(2)],
+//         // );
+
+//         Ok(())
+//     }
+
+//     #[cfg(all(feature = "async", target_os = "linux"))]
+//     #[tokio::test]
+//     async fn test_func_async_closure() -> Result<(), Box<dyn std::error::Error>> {
+//         {
+//             #[derive(Debug)]
+//             struct Data<T, S> {
+//                 _x: i32,
+//                 _y: String,
+//                 _v: Vec<T>,
+//                 _s: Vec<S>,
+//             }
+//             impl<T, S> Drop for Data<T, S> {
+//                 fn drop(&mut self) {
+//                     println!("Dropping Data");
+//                 }
+//             }
+
+//             let data: Data<i32, &str> = Data {
+//                 _x: 12,
+//                 _y: "hello".to_string(),
+//                 _v: vec![1, 2, 3],
+//                 _s: vec!["macos", "linux", "windows"],
+//             };
+
+//             // define an async closure
+//             let c = |_frame: CallingFrame,
+//                      _args: Vec<WasmValue>,
+//                      data: *mut std::os::raw::c_void|
+//              -> Box<
+//                 (dyn std::future::Future<Output = Result<Vec<WasmValue>, HostFuncError>> + Send),
+//             > {
+//                 // let host_data = unsafe { &mut *(data as *mut Data<i32, &str>) };
+//                 let host_data = unsafe { Box::from_raw(data as *mut Data<i32, &str>) };
+
+//                 Box::new(async move {
+//                     for _ in 0..10 {
+//                         println!("[async hello] say hello");
+//                         tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+
+//                         println!("host_data: {:?}", host_data);
+//                     }
+
+//                     println!("[async hello] Done!");
+
+//                     Ok(vec![])
+//                 })
+//             };
+
+//             // create a FuncType
+//             let result = FuncType::create(vec![], vec![]);
+//             assert!(result.is_ok());
+//             let func_ty = result.unwrap();
+
+//             // create an async host function
+//             let result =
+//                 Function::create_async_func(&func_ty, Box::new(c), Some(Box::new(data)), 0);
+//             assert!(result.is_ok());
+//             let async_hello_func = result.unwrap();
+
+//             // create an Executor
+//             let result = Executor::create(None, None);
+//             assert!(result.is_ok());
+//             let mut executor = result.unwrap();
+//             assert!(!executor.inner.0.is_null());
+
+//             // create a Store
+//             let result = Store::create();
+//             assert!(result.is_ok());
+//             let mut store = result.unwrap();
+
+//             // create an AsyncWasiModule
+//             let result = AsyncWasiModule::create(Some(vec!["abc"]), Some(vec![("a", "1")]), None);
+//             assert!(result.is_ok());
+//             let async_wasi_module = result.unwrap();
+
+//             // register async_wasi module into the store
+//             let wasi_import = ImportObject::AsyncWasi(async_wasi_module);
+//             let result = executor.register_import_object(&mut store, &wasi_import);
+//             assert!(result.is_ok());
+
+//             // create an ImportModule
+//             let mut import = ImportModule::create::<NeverType>("extern", None)?;
+//             import.add_func("async_hello", async_hello_func);
+
+//             let extern_import = ImportObject::Import(import);
+//             executor.register_import_object(&mut store, &extern_import)?;
+
+//             let extern_instance = store.module("extern")?;
+//             let async_hello = extern_instance.get_func("async_hello")?;
+
+//             async fn tick() {
+//                 let mut i = 0;
+//                 loop {
+//                     println!("[tick] i={i}");
+//                     tokio::time::sleep(std::time::Duration::from_millis(500)).await;
+//                     i += 1;
+//                 }
+//             }
+//             tokio::spawn(tick());
+
+//             let async_state = AsyncState::new();
+//             let _ = executor
+//                 .call_func_async(&async_state, &async_hello, [])
+//                 .await?;
+//         }
+
+//         assert_eq!(ASYNC_HOST_FUNCS.read().len(), 0);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
+
+//         Ok(())
+//     }
+
+//     #[cfg(all(feature = "async", target_os = "linux"))]
+//     #[tokio::test]
+//     async fn test_func_async_func() -> Result<(), Box<dyn std::error::Error>> {
+//         {
+//             #[derive(Debug)]
+//             struct Data<T, S> {
+//                 _x: i32,
+//                 _y: String,
+//                 _v: Vec<T>,
+//                 _s: Vec<S>,
+//             }
+//             let data: Data<i32, &str> = Data {
+//                 _x: 12,
+//                 _y: "hello".to_string(),
+//                 _v: vec![1, 2, 3],
+//                 _s: vec!["macos", "linux", "windows"],
+//             };
+
+//             // define async host function
+//             fn f<T: core::fmt::Debug + Send + Sync + 'static>(
+//                 _frame: CallingFrame,
+//                 _args: Vec<WasmValue>,
+//                 data: *mut std::ffi::c_void,
+//             ) -> Box<(dyn std::future::Future<Output = Result<Vec<WasmValue>, HostFuncError>> + Send)>
+//             {
+//                 let data = unsafe { Box::from_raw(data as *mut T) };
+
+//                 Box::new(async move {
+//                     for _ in 0..10 {
+//                         println!("[async hello] say hello");
+//                         tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+//                         println!("host_data: {:?}", data);
+//                     }
+
+//                     println!("[async hello] Done!");
+
+//                     Ok(vec![])
+//                 })
+//             }
+
+//             // create a FuncType
+//             let result = FuncType::create(vec![], vec![]);
+//             assert!(result.is_ok());
+//             let func_ty = result.unwrap();
+
+//             // create an async host function
+//             let result = Function::create_async_func(
+//                 &func_ty,
+//                 Box::new(f::<Data<i32, &str>>),
+//                 Some(Box::new(data)),
+//                 0,
+//             );
+//             assert!(result.is_ok());
+//             let async_hello_func = result.unwrap();
+
+//             // create an Executor
+//             let result = Executor::create(None, None);
+//             assert!(result.is_ok());
+//             let mut executor = result.unwrap();
+//             assert!(!executor.inner.0.is_null());
+
+//             // create a Store
+//             let result = Store::create();
+//             assert!(result.is_ok());
+//             let mut store = result.unwrap();
+
+//             // create an AsyncWasiModule
+//             let result = AsyncWasiModule::create(Some(vec!["abc"]), Some(vec![("a", "1")]), None);
+//             assert!(result.is_ok());
+//             let async_wasi_module = result.unwrap();
+
+//             // register async_wasi module into the store
+//             let wasi_import = ImportObject::AsyncWasi(async_wasi_module);
+//             let result = executor.register_import_object(&mut store, &wasi_import);
+//             assert!(result.is_ok());
+
+//             // create an ImportModule
+//             let mut import = ImportModule::create::<NeverType>("extern", None)?;
+//             import.add_func("async_hello", async_hello_func);
+
+//             let extern_import = ImportObject::Import(import);
+//             executor.register_import_object(&mut store, &extern_import)?;
+
+//             let extern_instance = store.module("extern")?;
+//             let async_hello = extern_instance.get_func("async_hello")?;
+
+//             async fn tick() {
+//                 let mut i = 0;
+//                 loop {
+//                     println!("[tick] i={i}");
+//                     tokio::time::sleep(std::time::Duration::from_millis(500)).await;
+//                     i += 1;
+//                 }
+//             }
+//             tokio::spawn(tick());
+
+//             let async_state = AsyncState::new();
+//             let _ = executor
+//                 .call_func_async(&async_state, &async_hello, [])
+//                 .await?;
+//         }
+
+//         assert_eq!(ASYNC_HOST_FUNCS.read().len(), 0);
+//         assert_eq!(HOST_FUNC_FOOTPRINTS.lock().len(), 0);
+
+//         Ok(())
+//     }
+// }
