+++ crates/wasmedge-sys/src/lib.rs
@@ -98,9 +98,6 @@ pub use config::Config;
 pub use executor::Executor;
 #[doc(inline)]
 pub use frame::CallingFrame;
-#[cfg(all(feature = "async", target_os = "linux"))]
-#[doc(inline)]
-pub use instance::function::AsyncHostFn;
 #[cfg(not(feature = "async"))]
 #[doc(inline)]
 pub use instance::module::WasiModule;
